Add security jwt for rest api

+++ src/main/java/com/luongtx/oes/service/BaseService.java
@@ -4,9 +4,14 @@ import java.util.List;
 
 public interface BaseService<T> {
     List<T> findAll();
-    T findById(Integer id);
-    T getById(Integer id);
+
+    T findById(Long id);
+
+    T getById(Long id);
+
     void add(T t);
-    T deleteById(Integer id);
+
+    T deleteById(Long id);
+
     void update(T t);
 }
